pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "1.0.0"
description = "A FastAPI boilerplate for REST APIs using relational DBs."
authors = ["Justin Largo <justinlargo7@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.scripts]
start = "app:main"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.32"
uvicorn = "^0.30.6"
fastapi = "^0.112.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
httpx = "^0.19.0"
pytest-cov = "^3.0.0"
black = "*"
ruff = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term-missing"
testpaths = ["tests"]